Meus voos

Subset da base do pacote anyflights, somente com os voos do ano de 2019 com origem nos aeroportos de MIA, JFK, ATL, FLL, IAH e MCO

meus_voos

Format

Uma dataframe com 1018199 linhas e 14 variavies:

ano

ano do voo

mes

mes do voo

dia

dia do voo

data

data do voo

semana

semana do voo

hora_partida

horario de partida do voo

atraso_partida

diferenca de entre entre a saida real e a saida programanda

atrasou_partida

indica se houve ou não atraso na partida

atraso_chegada

diferenca de entre entre a chegada real e a chegada programanda

atrasou_chegada

indica se houve ou não atraso na chegada

companhia

empresa aerea que opera o voo

origem

aeroporto de origem

destino

aeroporto de destino

distancia

distancia entre origem e destino

Source

https://github.com/simonpcouch/anyflights